[DF-308] Fixed warnings
Former-commit-id: c515b6cb0129b4bf5ab7026c3ddc4953ba5016a3

+++ Tools/ResourceEditor/Classes/TexturePacker/PngImage.cpp
@@ -58,26 +58,26 @@ bool PngImageExt::Create(uint32 width, uint32 height)
 	return (internalData != 0);
 }
 
-void PngImageExt::DrawImage(int sx, int sy, PngImageExt * image, const Rect2i & srcRect)
+void PngImageExt::DrawImage(int32 sx, int32 sy, PngImageExt * image, const Rect2i & srcRect)
 {
     uint32 * destData32 = (uint32*)GetData();
 	uint32 * srcData32 = (uint32*)image->GetData();
 	
-	int rx, ry;
+	int32 rx, ry;
 	ry = sy;
-	for (int y = srcRect.y; y < srcRect.y + srcRect.dy; ++y)
+	for (int32 y = srcRect.y; y < srcRect.y + srcRect.dy; ++y)
 	{
 		rx = sx;
-		for (int x = srcRect.x; x < srcRect.x + srcRect.dx; ++x)
+		for (int32 x = srcRect.x; x < srcRect.x + srcRect.dx; ++x)
 		{
 			if ((rx) < 0)continue;
-			if ((rx) >= GetWidth())continue;
+			if ((rx) >= (int32)GetWidth())continue;
 			if ((ry) < 0)continue;
-			if ((ry) >= GetHeight())continue;
+			if ((ry) >= (int32)GetHeight())continue;
 			if (x < 0)continue;
-			if (x >= image->GetWidth())continue;
+			if (x >= (int32)image->GetWidth())continue;
 			if (y < 0)continue;
-			if (y >= image->GetHeight())continue;
+			if (y >= (int32)image->GetHeight())continue;
 			
 			destData32[(rx) + (ry) * GetWidth()] = srcData32[x + y * image->GetWidth()];
 			//printf("%04x ", srcData32[x + y * image->width]);
@@ -87,7 +87,7 @@ void PngImageExt::DrawImage(int sx, int sy, PngImageExt * image, const Rect2i & 
 	}
 }
 
-void PngImageExt::DrawImage(int sx, int sy, PngImageExt * image)
+void PngImageExt::DrawImage(int32 sx, int32 sy, PngImageExt * image)
 {
 	// printf("0x%08x 0x%08x %d %d\n", data, image->data, sx, sy);
 	
@@ -95,13 +95,13 @@ void PngImageExt::DrawImage(int sx, int sy, PngImageExt * image)
 	uint32 * srcData32 = (uint32*)image->GetData();
 
     
-    for (int y = 0; y < image->GetHeight(); ++y)
-		for (int x = 0; x < image->GetWidth(); ++x)
+    for (int32 y = 0; y < image->GetHeight(); ++y)
+		for (int32 x = 0; x < image->GetWidth(); ++x)
 		{
 			if ((sx + x) < 0)continue;
-			if ((sx + x) >= GetWidth())continue;
+			if ((sx + x) >= (int32)GetWidth())continue;
 			if ((sy + y) < 0)continue;
-			if ((sy + y) >= GetHeight())continue;
+			if ((sy + y) >= (int32)GetHeight())continue;
 			
 			uint32 srcRead  = srcData32[x + y * image->GetWidth()];
 			destData32[(sx + x) + (sy + y) * GetWidth()] = srcRead;
@@ -110,19 +110,19 @@ void PngImageExt::DrawImage(int sx, int sy, PngImageExt * image)
 }
 
 
-bool PngImageExt::IsHorzLineOpaque(int y)
+bool PngImageExt::IsHorzLineOpaque(int32 y)
 {
 	uint8 * line = GetData() + y * GetWidth() * 4;
-	for (int x = 0; x < GetWidth(); ++x)
+	for (int32 x = 0; x < GetWidth(); ++x)
 		if (line[x * 4 + 3] != 0)
 			return false;
 	return true;
 }
 
-bool PngImageExt::IsVertLineOpaque(int x)
+bool PngImageExt::IsVertLineOpaque(int32 x)
 {
 	uint8 * vertLine = GetData() + x * 4;
-	for (int x = 0; x < GetHeight(); ++x)
+	for (int32 x = 0; x < GetHeight(); ++x)
 	{
 		if (vertLine[3] != 0)
 			return false;
@@ -135,14 +135,14 @@ bool PngImageExt::IsVertLineOpaque(int x)
 void PngImageExt::FindNonOpaqueRect(Rect2i & rect)
 {
 	rect = Rect2i(0, 0, GetWidth(), GetHeight());
-	for (int y = 0; y < GetHeight(); ++y)
+	for (int32 y = 0; y < GetHeight(); ++y)
 		if (IsHorzLineOpaque(y))
 		{
 			rect.y++;
 			rect.dy--;
 		}else break;
 	
-	for (int x = 0; x < GetWidth(); ++x)
+	for (int32 x = 0; x < GetWidth(); ++x)
 		if (IsVertLineOpaque(x))
 		{
 			rect.x++;
@@ -156,11 +156,11 @@ void PngImageExt::FindNonOpaqueRect(Rect2i & rect)
 		return;
 	}
 	
-	for (int y = GetHeight() - 1; y >= 0; --y)
+	for (int32 y = GetHeight() - 1; y >= 0; --y)
 		if (IsHorzLineOpaque(y))rect.dy--;
 		else break;
 	
-	for (int x = GetWidth() - 1; x >= 0; --x)
+	for (int32 x = GetWidth() - 1; x >= 0; --x)
 		if (IsVertLineOpaque(x))rect.dx--;
 		else break;
 }
@@ -169,12 +169,12 @@ void PngImageExt::DrawRect(const Rect2i & rect, uint32 color)
 {
 	uint32 * destData32 = (uint32*)GetData();
 	
-	for (int i = 0; i < rect.dx; ++i)
+	for (int32 i = 0; i < rect.dx; ++i)
 	{
 		destData32[rect.y * GetWidth() + rect.x + i] = color;
 		destData32[(rect.y + rect.dy - 1) * GetWidth() + rect.x + i] = color;
 	}
-	for (int i = 0; i < rect.dy; ++i)
+	for (int32 i = 0; i < rect.dy; ++i)
 	{
 		destData32[(rect.y + i) * GetWidth() + rect.x] = color;
 		destData32[(rect.y + i) * GetWidth() + rect.x + rect.dx - 1] = color;
